refactor: utils and types
