+++ app/[locale]/[...slug]/page.tsx
@@ -2,6 +2,8 @@ import pick from "lodash.pick"
 import { notFound } from "next/navigation"
 import { getMessages, setRequestLocale } from "next-intl/server"
 
+import { SlugPageParams } from "@/lib/types"
+
 import I18nProvider from "@/components/I18nProvider"
 import mdComponents from "@/components/MdComponents"
 
@@ -23,7 +25,7 @@ const loadData = dataLoader([["gfissues", fetchGFIs]])
 export default async function Page({
   params,
 }: {
-  params: Promise<{ locale: string; slug: string[] }>
+  params: Promise<SlugPageParams>
 }) {
   const { locale, slug: slugArray } = await params
 
@@ -112,7 +114,7 @@ export const dynamicParams = false
 export async function generateMetadata({
   params,
 }: {
-  params: Promise<{ locale: string; slug: string[] }>
+  params: Promise<SlugPageParams>
 }) {
   const { locale, slug } = await params
