+++ src/lib/utils/md.ts
@@ -5,7 +5,7 @@ import { extname, join } from "path"
 import matter from "gray-matter"
 import readingTime from "reading-time"
 
-import type { Frontmatter, ITutorial, Skill } from "@/lib/types"
+import type { Frontmatter, ITutorial, Skill, SlugPageParams } from "@/lib/types"
 
 import { dateToString } from "@/lib/utils/date"
 
@@ -100,3 +100,12 @@ export const getTutorialsData = (locale: string): ITutorial[] => {
 
   return tutorialData
 }
+
+export const checkPathValidity = (
+  validPaths: SlugPageParams[],
+  { locale, slug: slugArray }: SlugPageParams
+): boolean =>
+  validPaths.some(
+    (path) =>
+      path.locale === locale && path.slug.join("/") === slugArray.join("/")
+  )
